fix: Mark certain structs as !Sync
Some objects in dlib are not Thread Safe, thus they should be marked as !Sync
to do this in stable Rust, you add a field with `UnsafeCell`.

If you don't need `UnsafeCell`, then just put it inside a `PhantomData`, this
tells the type system that your type should act as if it held an `UnsafeCell`,
but it doesn't actually have it.

When [`negative_impls`](https://github.com/rust-lang/rust/issues/68318) is
stabilized, we can use that instead.

I would suggest applying this patch then bumping the crate version to 0.3.1
and yanking 0.3.0 to prevent people from downloading the previous version.